+++ sdk/resourcemanager/maps/armmaps/client.go
@@ -45,7 +45,7 @@ func NewClient(subscriptionID string, credential azcore.TokenCredential, options
 
 // NewListOperationsPager - List operations available for the Maps Resource Provider
 //
-// Generated from API version 2023-06-01
+// Generated from API version 2024-07-01-preview
 //   - options - ClientListOperationsOptions contains the optional parameters for the Client.NewListOperationsPager method.
 func (client *Client) NewListOperationsPager(options *ClientListOperationsOptions) *runtime.Pager[ClientListOperationsResponse] {
 	return runtime.NewPager(runtime.PagingHandler[ClientListOperationsResponse]{
@@ -78,7 +78,7 @@ func (client *Client) listOperationsCreateRequest(ctx context.Context, options *
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-06-01")
+	reqQP.Set("api-version", "2024-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -95,7 +95,7 @@ func (client *Client) listOperationsHandleResponse(resp *http.Response) (ClientL
 
 // NewListSubscriptionOperationsPager - List operations available for the Maps Resource Provider
 //
-// Generated from API version 2023-06-01
+// Generated from API version 2024-07-01-preview
 //   - options - ClientListSubscriptionOperationsOptions contains the optional parameters for the Client.NewListSubscriptionOperationsPager
 //     method.
 func (client *Client) NewListSubscriptionOperationsPager(options *ClientListSubscriptionOperationsOptions) *runtime.Pager[ClientListSubscriptionOperationsResponse] {
@@ -133,7 +133,7 @@ func (client *Client) listSubscriptionOperationsCreateRequest(ctx context.Contex
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-06-01")
+	reqQP.Set("api-version", "2024-07-01-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil